# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FAD Biometrics Configuration & Results Console — API interna
  description: >
    API interna de la consola (`/api/*`). No debe confundirse con la API externa de FAD
    Biometrics By Steps, documentada en `docs/api-contracts.md`. Toda llamada a FAD pasa por el
    backend a través de `FadApiAdapter`; el frontend nunca contacta a FAD directamente.
  version: "0.1.0"
servers:
  - url: /api
security:
  - cookieAuth: []
...
paths:
...
  /users:
    get:
      tags: [users]
      summary: Lista usuarios (ADMIN, AUDITOR)
      responses:
        "200": { description: OK }
    post:
      tags: [users]
      summary: Crea un usuario (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
                role: { type: string, enum: [ADMIN, OPERATOR, AUDITOR] }
      responses:
        "201": { description: Creado }
  /users/{id}:
    patch:
      tags: [users]
      summary: Actualiza rol/estado/contraseña de un usuario (ADMIN)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Actualizado }
    delete:
      tags: [users]
      summary: Elimina un usuario (ADMIN, nunca al último ADMIN activo)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "204": { description: Eliminado }
...
  /templates/{id}:
    get:
      tags: [templates]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: OK } }
    put:
      tags: [templates]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: Actualizada } }
    delete:
      tags: [templates]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "204": { description: Eliminada } }
  /templates/{id}/clone:
    post:
      tags: [templates]
      summary: Clona una plantilla existente
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "201": { description: Clonada } }
...
  /webhooks/fad:
    post:
      tags: [webhooks]
      summary: Endpoint público consumido por FAD. Basic Auth opcional configurable por ambiente.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, event]
              properties:
                id: { type: string }
                idUser: { type: string }
                event: { type: string }
                creationDate: { type: string }
                data: {}
                retry: { type: integer }
                error: { type: string, nullable: true }
                idOriginal: { type: string, nullable: true }
      responses:
        "200": { description: "Recibido (idempotente): { received: true, status }" }
        "401": { description: Basic Auth inválido }
...
  /webhooks/{id}:
    get:
      tags: [webhooks]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: OK } }
...
  /response-views/{id}:
    get:
      tags: [response-views]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: OK } }
    put:
      tags: [response-views]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: Actualizada } }
    delete:
      tags: [response-views]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "204": { description: Eliminada } }
  /response-views/{id}/apply/{executionId}:
    get:
      tags: [response-views]
      summary: Proyecta una vista sobre una ejecución (motor sin eval, ver docs/architecture.md)
      parameters:
        - { $ref: "#/components/parameters/id" }
        - name: executionId
          in: path
          required: true
          schema: { type: string }
      responses: { "200": { description: OK } }
...
  /settings/{key}:
    put:
      tags: [settings]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses: { "200": { description: OK } }
    delete:
      tags: [settings]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses: { "204": { description: Eliminado } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: fad_console_session
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }